doc: improve the readability of the prefix variable
Currently the default value for prefix - indicating the current
directory - is an empty string. This means that if we do
"{prefix}/foo", we may end up pointing to the root directory which is
not what we want. Make "./" be the default value so that the example
becomes valid and add "/" where we couldn't do it before.

Reviewed-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Erik Schilling <erik@riscstar.com>
Link: https://lore.kernel.org/r/20250605-rust-docs-v2-2-883a0a3872c0@linaro.org
Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>