build: fix pkgconfig check and drop unneeded AS_IF() in configure
The problem that was worked around with the AS_IF() hack turned out to
be caused by not expanding PKG_PROG_PKG_CONFIG() before the first
expansion of PKG_CHECK_MODULES(). Moving it near the top of the file
fixes the problem. This allows us to make the tests section consistent
with the rest of the configure.ac file.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>