dbus: manager: don't try to export the same chip twice
It's possible to get a second bind event from udevd (for instance:
manually triggered with `udevadm trigger -c bind`) on the same GPIO
chip so it already existing in the hashmap may actually happen unlike
what the comment in daemon.c states. We must not try to export the same
chip twice as it will crash the gpio-manager on the subsequent
assertion.

Fixes: a5ab76da1e0a ("dbus: add the D-Bus daemon, command-line client and tests")
Reported-by: Alexander Sverdlin <alexander.sverdlin@siemens.com>
Closes: https://github.com/brgl/libgpiod/issues/161
Tested-by: Alexander Sverdlin <alexander.sverdlin@siemens.com>
Link: https://lore.kernel.org/r/20251114132957.33750-1-brgl@bgdev.pl
Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>